WeaponsCard moved to container folder
