+++ src/components/WeaponsCard/WeaponsCard.css
