+++ src/components/WeaponsCardWeaponStats.jsx
@@ -2,11 +2,9 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import WeaponDataRow from './WeaponDataRow';
 import ButtonDeleteX from '../helpers/buttons/ButtonDeleteX';
-
-import { buildArrayForGunTable } from '../helpers/componentHelpers';
-
-import './WeaponsCard.css';
 import WeaponsCardModifyWeapon from './WeaponsCardModifyFirearm';
+import { buildArrayForGunTable } from '../helpers/componentHelpers';
+import './WeaponsCard/WeaponsCard.css';
 
 class WeaponsCardWeaponStats extends Component {
     getRangeBrackets = (gunObj) => {
